fix: Get rid of read-only template ref warning
We use a template ref in NativeDialog to hold the dialog element.,
The ref used to have the same name as the variable it is stored to.
This (presumably) causes this problem, thus we just rename it.

This should have been fixed in Vue 3.5.1/3.5.2 and it generally is.
see https://github.com/vuejs/core/commit/bc63df01992fdbf0b6749ad234153725697ed896

However, when using v-native-dialog as a dependency,
this problem still occurs for unknown reasons.