feat: Add 'closedby' support #1
This is the better solution for controlling how the dialog can be
closed by the user. `closedby="any"` allows for light dismissal.
`closedby="none"` prevents all user-initiated closing of the dialog.
Our previous implementation could only prevent this once and on a second
close attempt the dialog would close anyway.

'closedby' replaces the 'preventEscape' prop, which is now
deprecated. That property will now set `closedby="none"` itself.

Because of limited browser support for the 'closedby' attribute
(especially Safari) we use the previous implementation
(`preventDefault()` on the native `cancel` event)
as a fallback for `closedby="none"`. This is widely supported.

implements
preventEscape - double escape problem
https://github.com/kkuegler/v-native-dialog/issues/1